Movido classe do pacote pojo para model
