+++ src/main/java/br/edu/ifpe/garanhuns/sg/controller/ConsultaController.java
@@ -6,7 +6,7 @@
 package br.edu.ifpe.garanhuns.sg.controller;
 
 import br.edu.ifpe.garanhuns.sg.model.dao.hibernate.ConsultaHibernate;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Consulta;
+import br.edu.ifpe.garanhuns.sg.model.Consulta;
 import br.edu.ifpe.garanhuns.sg.model.pojo.enumarador.Especialidade;
 import br.edu.ifpe.garanhuns.sg.model.pojo.enumarador.Prioridade;
 import br.edu.ifpe.garanhuns.sg.model.pojo.enumarador.Status;
