+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/Atendente.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package br.edu.ifpe.garanhuns.sg.model.pojo;
+package br.edu.ifpe.garanhuns.sg.model;
 
 import java.io.Serializable;
 import javax.persistence.Entity;
