+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/hibernate/AtendenteHibernate.java
@@ -7,8 +7,8 @@ package br.edu.ifpe.garanhuns.sg.model.dao.hibernate;
 
 import java.util.List;
 import br.edu.ifpe.garanhuns.sg.model.dao.interfaces.AtendenteDAO;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Atendente;
-import br.edu.ifpe.garanhuns.sg.model.pojo.PostoSaude;
+import br.edu.ifpe.garanhuns.sg.model.Atendente;
+import br.edu.ifpe.garanhuns.sg.model.PostoSaude;
 import org.hibernate.Session;
 import br.edu.ifpe.garanhuns.sg.util.HibernateUtil;
