+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/hibernate/AtendimentoHibernate.java
@@ -7,8 +7,8 @@ package br.edu.ifpe.garanhuns.sg.model.dao.hibernate;
 
 import java.util.List;
 import br.edu.ifpe.garanhuns.sg.model.dao.interfaces.AtendimentoDAO;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Atendimento;
-import br.edu.ifpe.garanhuns.sg.model.pojo.PostoSaude;
+import br.edu.ifpe.garanhuns.sg.model.Atendimento;
+import br.edu.ifpe.garanhuns.sg.model.PostoSaude;
 import org.hibernate.Session;
 import br.edu.ifpe.garanhuns.sg.util.HibernateUtil;
