+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/hibernate/ConsultaHibernate.java
@@ -8,8 +8,8 @@ package br.edu.ifpe.garanhuns.sg.model.dao.hibernate;
 import java.util.ArrayList;
 import java.util.List;
 import br.edu.ifpe.garanhuns.sg.model.dao.interfaces.ConsultaDAO;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Consulta;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Paciente;
+import br.edu.ifpe.garanhuns.sg.model.Consulta;
+import br.edu.ifpe.garanhuns.sg.model.Paciente;
 import br.edu.ifpe.garanhuns.sg.model.pojo.enumarador.Status;
 import org.hibernate.Session;
 import br.edu.ifpe.garanhuns.sg.util.HibernateUtil;
