+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/hibernate/EnderecoHibernate.java
@@ -9,9 +9,9 @@ import java.util.List;
 import org.hibernate.Session;
 import br.edu.ifpe.garanhuns.sg.util.HibernateUtil;
 import br.edu.ifpe.garanhuns.sg.model.dao.interfaces.EnderecoDAO;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Bairro;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Paciente;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Endereco;
+import br.edu.ifpe.garanhuns.sg.model.Bairro;
+import br.edu.ifpe.garanhuns.sg.model.Paciente;
+import br.edu.ifpe.garanhuns.sg.model.Endereco;
 
 /**
  *
