+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/interfaces/ConsultaDAO.java
@@ -6,8 +6,8 @@
 package br.edu.ifpe.garanhuns.sg.model.dao.interfaces;
 
 import java.util.List;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Consulta;
-import br.edu.ifpe.garanhuns.sg.model.pojo.Paciente;
+import br.edu.ifpe.garanhuns.sg.model.Consulta;
+import br.edu.ifpe.garanhuns.sg.model.Paciente;
 
 /**
  *
