+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/interfaces/EnderecoDAO.java
@@ -5,7 +5,7 @@
  */
 package br.edu.ifpe.garanhuns.sg.model.dao.interfaces;
 
-import br.edu.ifpe.garanhuns.sg.model.pojo.Endereco;
+import br.edu.ifpe.garanhuns.sg.model.Endereco;
 
 /**
  *
