+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/interfaces/FilaAtandimentoDAO.java
@@ -5,7 +5,7 @@
  */
 package br.edu.ifpe.garanhuns.sg.model.dao.interfaces;
 
-import br.edu.ifpe.garanhuns.sg.model.pojo.FilaAtendimento;
+import br.edu.ifpe.garanhuns.sg.model.FilaAtendimento;
 
 /**
  *
